chore(dep): remove global dependencies for dev dependencies
Gulp and Bower shouldn't be installed globally, because we are all working on different project with different needs and don't want to have a new version of X or Y required for this project. To avoid this Prerequisites, we can install them as dev-dep in package.json and aliasing them in scripts. The result is event simpler than befoe (with pre install hook for example).
BTW, all the docker config was totally broken due to the use of npm and local dependencies... The installation was done during the build process but the shared volume overide it and the node_modules, bower_components and other folder was missing during serve. So I fix the docker system and modify the documentation.
For now, you can install dependencies, serve and build. I will try to allow the firebase deployment operation later. By default, docker container do installation and serving if no parameter override the CMD (see the doc).