fix(reload): error during reloading base on markdown and images
While running gulp serve, when a file into `/data/**` or an image was changed, an error happened in the console and the server totally stop itself. 
This was because the `.on` method of `gulp.watch` doesn't give a callback but only the filename, so calling `done()` is first useless and end in error because the parameter was a string and not a function.